// SourceCode/Application/VedSoft/VedSoft.Angular/src/app/core/models/master-model/customer-model.ts
export class CustomerModel
    {
        customerId?: number;
        name?: string;
        code?: string;
        subDomain?: string;
        description?: string;
        active?: number;
        contactNumber?: string;
        address?: string;
        otherInfo?: string;
        createdDate?: Date | string;
        createdBy?: number;
    }